refractor torrent to make it easier to add new sources
moved the storage creation into a module ‘torrent-manager’ this will
allow for further source to be added easily.
