+++ lib/torrent.js
@@ -14,11 +14,11 @@ var parseTorrent = require('parse-torrent')
 var RarityMap = require('./rarity-map')
 var reemit = require('re-emitter')
 var Server = require('./server') // browser exclude
-var Storage = require('./storage')
 var Swarm = require('bittorrent-swarm') // `webtorrent-swarm` in browser
 var url = require('url')
 var ut_metadata = require('ut_metadata')
 var ut_pex = require('ut_pex') // browser exclude
+var Storage = require('./storage')
 
 var MAX_BLOCK_LENGTH = 128 * 1024
 var MAX_OUTSTANDING_REQUESTS = 5
@@ -39,16 +39,15 @@ inherits(Torrent, EventEmitter)
  * @param {string|Buffer|Object} torrentId
  * @param {Object} opts
  */
-function Torrent (torrentId, opts) {
+function Torrent (manager, parsedTorrent, opts) {
   var self = this
   EventEmitter.call(self)
   debug('new torrent')
 
   self.client = opts.client
-
+  self.manager = manager;
   self.hotswapEnabled = ('hotswap' in opts ? opts.hotswap : true)
   self.verify = opts.verify
-  self.storageOpts = opts.storageOpts
 
   self.chokeTimeout = opts.chokeTimeout || CHOKE_TIMEOUT
   self.pieceTimeout = opts.pieceTimeout || PIECE_TIMEOUT
@@ -62,64 +61,37 @@ function Torrent (torrentId, opts) {
   self.ready = false
   self.files = []
   self.metadata = null
-  self.parsedTorrent = null
+  self.parsedTorrent = parsedTorrent
+  self.infoHash = parsedTorrent.infoHash
+  if (parsedTorrent.name) self.name = parsedTorrent.name // preliminary name
   self.storage = null
   self.numBlockedPeers = 0
   self._amInterested = false
   self._destroyed = false
   self._selections = []
   self._critical = []
-  self._storageImpl = opts.storage || Storage
+ 
 
-  var parsedTorrent = parseTorrent(torrentId)
-  if (parsedTorrent && parsedTorrent.infoHash) {
-    onTorrentId(parsedTorrent)
+  // create swarm
+  self.swarm = new Swarm(self.infoHash, self.client.peerId, {
+    handshake: { dht: !!self.client.dht }
+  })
+  reemit(self.swarm, self, ['warning', 'error'])
+  self.swarm.on('wire', self._onWire.bind(self))
 
-  } else if (typeof hh.get === 'function' && /^https?:/.test(torrentId)) {
-    // http or https url to torrent file
-    httpGet(torrentId, function (err, torrent) {
-      if (err)
-        return self.emit('error', new Error('error downloading torrent: ' + err.message))
-      onTorrentId(torrent)
-    })
+  // update overall client stats
+  self.swarm.on('download', self.client.downloadSpeed.bind(self.client))
+  self.swarm.on('upload', self.client.uploadSpeed.bind(self.client))
 
-  } else if (typeof fs.readFile === 'function') {
-    // assume it's a filesystem path
-    fs.readFile(torrentId, function (err, torrent) {
-      if (err) return self.emit('error', new Error('invalid torrent id'))
-      onTorrentId(torrent)
-    })
-
-  } else throw new Error('invalid torrent id')
-
-  function onTorrentId (torrentId) {
-    parsedTorrent = parseTorrent(torrentId)
-    self.infoHash = parsedTorrent.infoHash
-    if (parsedTorrent.name) self.name = parsedTorrent.name // preliminary name
-
-    // create swarm
-    self.swarm = new Swarm(self.infoHash, self.client.peerId, {
-      handshake: { dht: !!self.client.dht }
-    })
-    reemit(self.swarm, self, ['warning', 'error'])
-    self.swarm.on('wire', self._onWire.bind(self))
-
-    // update overall client stats
-    self.swarm.on('download', self.client.downloadSpeed.bind(self.client))
-    self.swarm.on('upload', self.client.uploadSpeed.bind(self.client))
-
-    if (process.browser) {
-      // in browser, swarm does not listen
-      self._onSwarmListening(parsedTorrent)
-    } else {
-      // listen for peers
-      self.swarm.listen(self.client.torrentPort, self._onSwarmListening.bind(self, parsedTorrent))
-    }
-    process.nextTick(function () {
-      self.emit('infoHash')
-    })
+  if (process.browser) {
+    // in browser, swarm does not listen
+    self._onSwarmListening(parsedTorrent)
+  } else {
+    // listen for peers
+    self.swarm.listen(self.client.torrentPort, self._onSwarmListening.bind(self, parsedTorrent))
   }
 }
+
 
 // torrent size (in bytes)
 Object.defineProperty(Torrent.prototype, 'length', {
@@ -164,62 +136,10 @@ Object.defineProperty(Torrent.prototype, 'ratio', {
   }
 })
 
-Torrent.prototype._onSwarmListening = function (parsed, port) {
+Torrent.prototype.attachStorage = function(storage) {
   var self = this
-  if (self._destroyed) return
-
-  self.client.torrentPort = port
-
-  // begin discovering peers via the DHT and tracker servers
-  self.discovery = new Discovery({
-    announce: parsed.announce,
-    dht: self.client.dht,
-    tracker: self.client.tracker,
-    peerId: self.client.peerId,
-    port: port
-  })
-  self.discovery.setTorrent(self.infoHash)
-  self.discovery.on('peer', self.addPeer.bind(self))
-
-  // expose discovery events
-  reemit(self.discovery, self, ['dhtAnnounce', 'warning', 'error'])
-
-  // if full metadata was included in initial torrent id, use it
-  if (parsed.info) self._onMetadata(parsed)
-
-  self.emit('listening', port)
-}
-
-/**
- * Called when the metadata is received.
- */
-Torrent.prototype._onMetadata = function (metadata) {
-  var self = this
-  if (self.metadata || self._destroyed) return
-  debug('got metadata')
-
-  if (metadata && metadata.infoHash) {
-    // `metadata` is a parsed torrent (from parse-torrent module)
-    self.metadata = parseTorrent.toBuffer(metadata)
-    self.parsedTorrent = metadata
-  } else {
-    self.metadata = metadata
-    try {
-      self.parsedTorrent = parseTorrent(self.metadata)
-    } catch (err) {
-      return self.emit('error', err)
-    }
-  }
-
-  // update preliminary torrent name
-  self.name = self.parsedTorrent.name
-
-  // update discovery module with full torrent metadata
-  self.discovery.setTorrent(self.parsedTorrent)
-
-  self.rarityMap = new RarityMap(self.swarm, self.parsedTorrent.pieces.length)
-
-  self.storage = new self._storageImpl(self.parsedTorrent, self.storageOpts)
+  self.storage = storage
+  debug('attaching storage')
   self.storage.on('piece', self._onStoragePiece.bind(self))
   self.storage.on('file', function (file) {
     self.emit('file', file)
@@ -244,6 +164,21 @@ Torrent.prototype._onMetadata = function (metadata) {
   self.storage.files.forEach(function (file) {
     self.files.push(file)
   })
+  
+  process.nextTick(self._onStorage.bind(self))
+}
+
+Torrent.prototype.updateMetadata = function(parsedTorrent) {
+  var self = this
+  self.parsedTorrent = parsedTorrent
+  // update preliminary torrent name
+  self.name = self.parsedTorrent.name
+  self.metadata = parseTorrent.toBuffer(parsedTorrent)
+
+  // update discovery module with full torrent metadata
+  self.discovery.setTorrent(self.parsedTorrent)
+
+  self.rarityMap = new RarityMap(self.swarm, self.parsedTorrent.pieces.length)
 
   self.swarm.wires.forEach(function (wire) {
     // If we didn't have the metadata at the time ut_metadata was initialized for this
@@ -252,43 +187,31 @@ Torrent.prototype._onMetadata = function (metadata) {
 
     self._onWireWithMetadata(wire)
   })
-
-  if (self.verify) {
-    process.nextTick(function () {
-      debug('verifying existing torrent data')
-      var numPieces = 0
-      var numVerified = 0
-
-      // TODO: move storage verification to storage.js?
-      parallel(self.storage.pieces.map(function (piece) {
-        return function (cb) {
-          self.storage.read(piece.index, function (err, buffer) {
-            numPieces += 1
-            self.emit('verifying', {
-              percentDone: 100 * numPieces / self.storage.pieces.length,
-              percentVerified: 100 * numVerified / self.storage.pieces.length,
-            })
-
-            if (!err && buffer) {
-              // TODO: this is a bit hacky; figure out a cleaner way of verifying the buffer
-              piece.verify(buffer)
-              numVerified += piece.verified
-              debug('piece ' + (piece.verified ? 'verified' : 'invalid') + ' ' + piece.index)
-            }
-            // continue regardless of whether piece verification failed
-            cb()
-          }, true) // forces override to allow reading from non-verified pieces
-        }
-      }), self._onStorage.bind(self))
-    })
-  } else {
-    process.nextTick(self._onStorage.bind(self))
-  }
-
-  process.nextTick(function () {
-    self.emit('metadata')
-  })
 }
+
+Torrent.prototype._onSwarmListening = function (parsed, port) {
+  var self = this
+  if (self._destroyed) return
+
+  self.client.torrentPort = port
+
+  // begin discovering peers via the DHT and tracker servers
+  self.discovery = new Discovery({
+    announce: parsed.announce,
+    dht: self.client.dht,
+    tracker: self.client.tracker,
+    peerId: self.client.peerId,
+    port: port
+  })
+  self.discovery.setTorrent(self.infoHash)
+  self.discovery.on('peer', self.addPeer.bind(self))
+
+  // expose discovery events
+  reemit(self.discovery, self, ['dhtAnnounce', 'warning', 'error'])
+
+  self.emit('listening', port)
+}
+
 
 /**
  * Destroy and cleanup this torrent.
@@ -403,14 +326,14 @@ Torrent.prototype.critical = function (start, end) {
 
 Torrent.prototype._onWire = function (wire) {
   var self = this
-
+  debug(!!self.metadata ? "Metadata" : " No MEta");
   // use ut_metadata extension
   wire.use(ut_metadata(self.metadata))
 
   if (!self.metadata) {
     wire.ut_metadata.on('metadata', function (metadata) {
       debug('got metadata via ut_metadata')
-      self._onMetadata(metadata)
+      self.manager.onMetadata(metadata)
     })
     wire.ut_metadata.fetch()
   }
@@ -963,13 +886,6 @@ Torrent.prototype._request = function (wire, index, hotswap) {
   wire.request(index, block.offset, block.length, gotPiece)
 
   return true
-}
-
-Torrent.prototype.createServer = function (opts) {
-  var self = this
-  if (typeof Server === 'function' /* browser exclude */) {
-    return new Server(self, opts)
-  }
 }
 
 /**
